pyproject.toml
[build-system]
requires = ["poetry>=0.12", "setuptools>=36.2.2", "wheel>=0.28.0"]
build-backend = "poetry.masonry.api"

[tool.poetry]
name = "ffmeta"
version = "0.1.0"
description = "Tool to write media metadata using ffmpeg."
authors = ["st37 <st37@tuta.io>"]
maintainers = []
license = "ISC"
packages = [
  { include = "ffmeta", from = "src" }
]
readme = "README.md"
homepage = "https://github.com/stephen-bunn/ffmeta"
repository = "https://github.com/stephen-bunn/ffmeta"
documentation = "https://ffmeta.readthedocs.io"
keywords = [
  "ffmeta"
]
classifiers = [
  "License :: OSI Approved :: ISC License (ISCL)",
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Operating System :: POSIX",
  "Operating System :: MacOS :: MacOS X",
  "Operating System :: Microsoft :: Windows",
  "Natural Language :: English",
  "Topic :: Software Development :: Libraries :: Python Modules"
]
include = [
  "pyproject.toml",
  "LICENSE",
  "README*.rst",
  "README*.md",
  "CHANGELOG.rst",
  "CHANGELOG.md",
  "CONTRIBUTING.rst",
  "CONTRIBUTING.md",
]
exclude = [
  ".editorconfig",
  ".pre-commit-config.yaml",
  ".readthedocs.yaml",
  "requirements*.txt"
]

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/stephen-bunn/ffmeta/issues"

[tool.poetry.dependencies]
python = "^3.10.0"
rich = "^10.14.0"
typer = "^0.4.0"
ffmpeg-python = "^0.2.0"
wrapt = "^1.13.3"
attrs = "^21.2.0"
python-magic = "^0.4.24"

[tool.poetry.dev-dependencies]
black = "^21.11b1"
coverage = { version = "^5.2.1", extras = ["toml"] }
flake8 = "^3.8.3"
flake8-docstrings = "^1.5.0"
hypothesis = "^5.29.3"
isort = "^5.4.2"
mypy = "^0.782"
pre-commit = "^2.7.1"
pytest = "^6.0.1"
pytest-cov = "^2.10.1"
pytest-flake8 = "^1.0.6"
pytest-sugar = "^0.9.4"
pytest-xdist = "^2.1.0"
rope = "^0.17.0"
toml = "^0.10.1"

[tool.poetry.scripts]
ffmeta = "ffmeta.cli:app"

[tool.black]
line-length = 88
target_version = ['py37']
include = '\.pyi?$'
exclude = '''
(
  /(
    \.eggs
    | \.git
    | \.mypy_cache
    | _build
    | build
    | dist
  )/
)
'''

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-rxsX --flake8 -n 4 --cov"
norecursedirs = [
  ".git",
  "_build",
  "dist",
  "changes",
  "tasks",
  "docs"
]
testpaths = "tests"
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
flake8-ignore = [
  "docs/source/*.py ALL",
  "tests/*.py ALL",
  "setup.py ALL"
]

[tool.coverage.run]
branch = true
parallel = true
disable_warnings = [
  "no-data-collected"
]
source = [
  "src/ffmeta/"
]

[tool.coverage.report]
show_missing = true

[tool.coverage.html]
directory = "htmlcov"
